update routing + store
